test/reftest/csswg-test/Converted/css-transforms-1/transform-2d-getComputedStyle-001.html

<!DOCTYPE html>
<html><head>
    <title>CSS Transforms API Test: transform translate</title>
    <link rel="author" title="Mihai Balan" href="mailto:mibalan@adobe.com">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="CSS 2D transforms correctly report their matrix via getComputedStyle()">
    
    <script id="metadata_cache">/*
	{
	  "Matrix for translation transforms": {},
	  "Matrix for rotate": {},
	  "Matrix for scaling": {},
	  "Matrix for skew": {},
	  "Matrix for general transform": {}
	}
	*/</script>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
<style>.convertedClass1 {  
  display: block;  
  width: 50px;  
  height: 50px;  
  background-color: green;
}
.convertedClass2 {  
  transform: translate(10px, 20px);  
  -webkit-transform: translate(10px, 20px);
}
.convertedClass3 {  
  transform: translateX(10px);  
  -webkit-transform: translateX(10px);
}
.convertedClass4 {  
  transform: translateY(20px);  
  -webkit-transform: translateY(20px);
}
.convertedClass5 {  
  transform: rotate(90deg);  
  -webkit-transform: rotate(90deg);
}
.convertedClass6 {  
  transform: scale(2.0);  
  -webkit-transform: scale(2.0);
}
.convertedClass7 {  
  transform: scaleX(0.5);  
  -webkit-transform: scaleX(0.5);
}
.convertedClass8 {  
  transform: scaleY(1.5);  
  -webkit-transform: scaleY(1.5);
}
.convertedClass9 {  
  transform: skewX(45deg);  
  -webkit-transform: skewX(45deg);
}
.convertedClass10 {  
  transform: skewY(45deg);  
  -webkit-transform: skewY(45deg);
}
.convertedClass11 {  
  transform: matrix(1, 2, 3, 4, 5, 6);  
  -webkit-transform: matrix(1, 2, 3, 4, 5, 6);
}
</style></head>
<body>
	<div id="log"></div>

	<div id="translate" class="block convertedClass1 convertedClass2"></div>
	<div id="translateX" class="block convertedClass1 convertedClass3"></div>
	<div id="translateY" class="block convertedClass1 convertedClass4"></div>
	<div id="rotate" class="block convertedClass1 convertedClass5"></div>
	<div id="scale" class="block convertedClass1 convertedClass6"></div>
	<div id="scaleX" class="block convertedClass1 convertedClass7"></div>
	<div id="scaleY" class="block convertedClass1 convertedClass8"></div>
	<div id="skewX" class="block convertedClass1 convertedClass9"></div>
	<div id="skewY" class="block convertedClass1 convertedClass10"></div>
	<div id="matrix" class="block convertedClass1 convertedClass11"></div>
    <script type="text/javascript">
    function getTransformFor(id) {
    	return window.getComputedStyle(document.getElementById(id)).getPropertyValue("transform");
    }
    function clear(id) {
    	document.getElementById(id).style.display = 'none';
    }

    test(function() {
    	assert_equals(getTransformFor("translate"), "matrix(1, 0, 0, 1, 10, 20)", "Incorrect matrix for translate()");
    	clear("translate");
    	assert_equals(getTransformFor("translateX"), "matrix(1, 0, 0, 1, 10, 0)", "Incorrect matrix for translateX()");
    	clear("translateX");
    	assert_equals(getTransformFor("translateY"), "matrix(1, 0, 0, 1, 0, 20)", "Incorrect matrix for translateY()");
    	clear("translateY");
    }, "Matrix for translation transforms");

    test(function() {
    	assert_equals(getTransformFor("rotate"), "matrix(0, 1, -1, 0, 0, 0)", "Incorrect matrix for rotate()");
    	clear("rotate");
    }, "Matrix for rotate");

    test(function() {
    	assert_equals(getTransformFor("scale"), "matrix(2, 0, 0, 2, 0, 0)", "Incorrect matrix for scale()");
    	clear("scale");
    	assert_equals(getTransformFor("scaleX"), "matrix(0.5, 0, 0, 1, 0, 0)", "Incorrect matrix for scaleX()");
    	clear("scaleX");
    	assert_equals(getTransformFor("scaleY"), "matrix(1, 0, 0, 1.5, 0, 0)", "Incorrect matrix for scaleY()");
    	clear("scaleY");
    }, "Matrix for scaling");

    test(function() {
    	assert_equals(getTransformFor("skewX"), "matrix(1, 0, 1, 1, 0, 0)", "Incorrect matrix for skewX()");
    	clear("skewX");
    	assert_equals(getTransformFor("skewY"), "matrix(1, 1, 0, 1, 0, 0)", "Incorrect matrix for skewY()");
    	clear("skewY");
    }, "Matrix for skew");

    test(function() {
    	assert_equals(getTransformFor("matrix"), "matrix(1, 2, 3, 4, 5, 6)", "Incorrect matrix for matrix()");
    	clear("matrix");
    }, "Matrix for general transform");
    </script>


</body></html>
